test: add test-whitelist/ folder
