+++ test-whitelist/test.wpcom.me.settings.password.js
