--- test/test.wpcom.me.connected-application.js
@@ -1,46 +1,0 @@
-
-/**
- * Module dependencies
- */
-var util = require( './util' );
-var assert = require( 'assert' );
-
-/**
- * me.connectedApp
- */
-describe( 'wpcom.me.connectedApp', function() {
-	// Global instances
-	var wpcom = util.wpcom();
-	var me = wpcom.me();
-
-	describe( 'wpcom.me.connectedApp', function() {
-		it( 'should get current user\' connected applications', function( done ) {
-			me.connectedApps( function( err, data ) {
-				if ( err ) throw err;
-
-				assert.ok( data );
-				assert.ok( data.connected_applications );
-
-				let appId = data.connected_applications && data.connected_applications[0]
-					? data.connected_applications[0].ID
-					: null;
-
-				if ( appId ) {
-					describe( 'wpcom.me.connectedApp.get', function() {
-						it( 'should get current user\' connected applications',
-						function( done2 ) {
-							me.connectedApp( appId ).get( function( err2, data2 ) {
-								if ( err2 ) throw err2;
-
-								assert.ok( data2 );
-								assert.equal( appId, data2.ID );
-								done2();
-							} );
-						} );
-					} );
-				}
-				done();
-			} );
-		} );
-	} );
-} );
