+++ test/test.wpcom.me.js
@@ -96,18 +96,6 @@ describe( 'wpcom.me', function() {
 		} );
 	} );
 
-	describe( 'wpcom.me.connectedApps', function() {
-		it( 'should get current user\' connected applications', done => {
-			me.connectedApps()
-				.then( data => {
-					assert.ok( data );
-					assert.ok( data.connected_applications instanceof Array );
-					done();
-				} )
-				.catch( done );
-		} );
-	} );
-
 	describe( 'wpcom.me.keyringConnections', function() {
 		it( 'should get current user\' keyring connections', done => {
 			me.keyringConnections()
